// _REPO_BAUL_SEGURIDAD/backups/backup_20260206_2211/fidelidad-next/src/types.ts
export interface Prize {
    id: string;
    name: string;
    pointsRequired: number;
    stock: number;
    description?: string;
    active: boolean;
    imageUrl?: string;
    cashValue?: number; // Valor en pesos para reportes
}